Fixed Sample
Added missing DXUT
Added missing IntelPowerGadgetLib.h/IntelPowerGadgetLib.cpp
Upgraded solution file to vs2015
changed HueToRGB - changed modf to modff
Linked legacy_stdio_definitions.lib to vs2015

+++ ColorUtil.cpp
@@ -17,7 +17,7 @@
 D3DXVECTOR3 HueToRGB(float hue)
 {
     float intPart;
-    float fracPart = modf(hue * 6.0f, &intPart);
+    float fracPart = modff(hue * 6.0f, &intPart);
     int region = static_cast<int>(intPart);
     
     switch (region) {
